Updated BlockParty
Updated .gitignore to not include .gitignore (it automatically omits it anyways, but for the sake of professionality).
Fixed Start subcommand to not display the host's name as CraftPlayer(Name=~~~). Not sure how that snuck in there. There might still be a few stragglers here and there.
When the plugin is unloaded, all the events are removed.
Fixed BlockParty giving an error when attempting to end it before the runnable started.
BlockParty now removes blocks to allow players to fall instead of killing them directly.
BlockParty displays the current color in the subtitle, and also the bossbar.
BlockParty now fully utilizes the bossbar to display time remaining and current color.
BlockParty now displays subtitle/bossbar/sounds to event host even if they're not playing.